Enhance TextRegionAccessBuildingSequencer to fix Quickfix formatting
Prevent undesired newlines being inserted. Two existing Xtext tests
needed to be adapted as well.

When the formatter runs in "Quickfix mode" it is set to
"formatUndefinedHiddenRegionsOnly". The trailing hidden region following
the to be formatted EObjectRegion was set by the
TextRegionAccessBuildingSequencer to "undefined" which caused the text
replacement causing the undesired newlines being included in the text
replacements created by the formatter.
To prevent this, the TextRegionAccessBuildingSequencer now sets the
trailing hidden region to "defined".

Signed-off-by: Bernd Lehmann <bernd.lehmann@posteo.de>